Use UNSPECIFIED as the default socket address
The default config of the DHT insntace used a localhost-only binding
socket by default. It means that when the service boostraps by default,
it cannot communicates with the external internet.

The following default was swalling errors from the socket:

```rust
let mut state = HyperDht::with_config(DhtConfig::default())
	            .await
		    .unwrap();
// Swalloed error
// Err(Custom { kind: Other, error: VerboseError { source: Os { code: 10051, kind: Other, message: "A socket operation was attempted to an unreachable network." }, message: "could not send packet to 35.209.238.143:49737" } })
state.next().await;

```

This error was mitigated by setting a manual socket, but the default
should allow connections to the main hyperswarm DHT nodes.

This commit changes the default to use `0.0.0.0` instead of `127.0.0.1`
to allow internet connections on the socket.